Move media viewer modal into evidence section
