+++ src/i18n/l10n/en.ftl.json
@@ -232,6 +232,7 @@
   },
   "No-comments-or-ids-to-display": "No comments or ids to display",
   "No-Location": "No Location",
+  "none": "none",
   "No-photos-found": "No photos found. If this is your first time opening the app and giving permissions, try restarting the app.",
   "no-rights-reserved": {
     "comment": "license code",
@@ -660,5 +661,13 @@
   "The-location-will-not-be-visible": "The location will not be visible to others, which means it may be difficult to identify.",
   "#Wild status sheet descriptions": "",
   "This-is-a-wild-organism": "This is a wild organism and wasn’t placed in this location by humans.",
-  "This-organism-was-placed-by-humans": "This organism was placed in this location by humans. This applies to things like garden plants, pets, and zoo animals."
+  "This-organism-was-placed-by-humans": "This organism was placed in this location by humans. This applies to things like garden plants, pets, and zoo animals.",
+  "Lat-Lon-Acc": {
+    "comment": "Latitude, longitude, and accuracy",
+    "val": "Lat: { $latitude }, Lon: { $longitude }, Acc: { $accuracy }"
+  },
+  "Every-observation-needs": {
+    "comment": "Missing evidence sheet",
+    "val": "Every observation needs a location, date, and time to be helpful to identifiers. You can edit geoprivacy if you’re concerned about location privacy."
+  }
 }
